+++ CHANGELOG.md
@@ -53,6 +53,7 @@
   - improve model cpu offload compatibility
   - improve model sequential offload compatibility
   - improve bfloat16 compatibility
+  - improve xformers installer to match cuda version and install triton
   - fix extra networks refresh
   - fix sdp memory attention in backend original
   - fix autodetect sd21 models